# feat: convert `PrintCommand` to value type