--- a/Src/CSharpier/DocPrinter/DocPrinter.cs
+++ b/Src/CSharpier/DocPrinter/DocPrinter.cs
@@ -405,7 +405,7 @@ internal class DocPrinter
     }
 }
 
-internal record PrintCommand(Indent Indent, PrintMode Mode, Doc Doc);
+internal record struct PrintCommand(Indent Indent, PrintMode Mode, Doc Doc);
 
 internal enum PrintMode
 {